requirements.txt
matplotlib==3.3.2
numpy==1.18.5
opencv-python==4.4.0.44
pdbpp==0.10.2
tensorboard==2.3.0
tensorflow-gpu==2.3.0
scipy==1.4.1
Pillow==8.0.1
gdown==3.13.0

